+++ docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @midl-xyz/midl-js-docs
 
+## 0.0.5
+
+### Patch Changes
+
+- Add publish param to signPSBT
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.4
+  - @midl-xyz/midl-js-react@0.0.6
+
 ## 0.0.4
 
 ### Patch Changes