+++ playground/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @midl-xyz/midl-js-playground
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.4
+  - @midl-xyz/midl-js-react@0.0.6
+
 ## 0.1.4
 
 ### Patch Changes